Refactor/rename entity and strategy (#76)
* Refactor: Clean up unused imports in FCM controller and module

* Feat/fcm: Fix push test endpoint

* Refactor: Rename entity files

* Refactor: Rename group.owner.strategy.ts to group-owner.strategy.ts

* Format

---------
