+++ src/auth/strategies/group-owner.strategy.ts
