+++ src/fcm/fcm.module.ts
@@ -3,7 +3,7 @@ import { TypeOrmModule } from '@nestjs/typeorm';
 
 import { FcmService } from './fcm.service';
 import { FcmController } from './fcm.controller';
-import { FcmKey } from './entities/fcm.key.entity';
+import { FcmKey } from './entities/fcm-key.entity';
 
 @Module({
   imports: [TypeOrmModule.forFeature([FcmKey])],
