+++ src/user/entities/user.entity.ts
@@ -11,9 +11,9 @@ import {
 import { Exclude } from 'class-transformer';
 import { ApiHideProperty } from '@nestjs/swagger';
 
-import { FcmKey } from 'src/fcm/entities/fcm.key.entity';
+import { FcmKey } from 'src/fcm/entities/fcm-key.entity';
 import { Room } from 'src/room/entities/room.entity';
-import { RoomUser } from 'src/room/entities/room.user.entity';
+import { RoomUser } from 'src/room/entities/room-user.entity';
 import { Chat } from 'src/chat/entities/chat.entity';
 import { Report } from 'src/report/entities/report.entity';
